requirements.txt
aiohttp==3.5.4
async-timeout==3.0.1
attrs==19.1.0
certifi==2019.3.9
chardet==3.0.4
Django==2.2
djangorestframework==3.9.2
idna==2.8
idna-ssl==1.1.0
jsonfield==2.0.2
multidict==4.5.2
Pillow==6.0.0
psycopg2-binary==2.8.2
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
typing-extensions==3.7.2
virtualenv==16.4.3
virtualenv-clone==0.5.3
yarl==1.3.0
